Refactor keycloak client
