+++ pkg/keycloak/client_service_test.go
@@ -1,0 +1,1 @@
+package keycloak
