+++ pkg/keycloak/client_test.go
@@ -13,7 +13,7 @@ import (
 func ExampleNewClient() {
 	u, _ := url.Parse("http://localhost/auth/admin")
 	c := NewClient(*u, http.DefaultClient)
-	userID, _ := c.Users.Create(context.Background(), "myRealm", &UserRepresentation{
+	userID, _ := c.Users().Create(context.Background(), "myRealm", &UserRepresentation{
 		Username: "hello-world",
 	})
 	fmt.Println("UserID: ", userID)
@@ -28,7 +28,6 @@ func TestNewClient(t *testing.T) {
 
 	a.NotNil(client)
 	a.NotNil(client)
-	a.NotNil(client.Users)
 	a.False(client.restClient.Debug)
 }
